Nuevas funciones

+++ sercor/Factura.cs
@@ -11,15 +11,15 @@ namespace sercor
         public int ID_FACTURA { get; set; }
         public string ID_CLIENTE { get; set; }
         public int ID_USUARIO { get; set; }
-        public float IVA { get; set; }
-        public float TOTAL { get; set; }
+        public decimal IVA { get; set; }
+        public decimal TOTAL { get; set; }
         public string FECHA { get; set; }
 
 
         //Constructor de objeto
         public Factura() { }
 
-        public Factura(int pIdFactura, string pIdCliente, int pIdUsuario, float pIva, float pTotal, string pFecha)
+        public Factura(int pIdFactura, string pIdCliente, int pIdUsuario, decimal pIva, decimal pTotal, string pFecha)
         {
             this.ID_FACTURA = pIdFactura;
             this.ID_CLIENTE = pIdCliente;
